Create README.md
INTRODUCTION
In the existing system the people must go to the office, or any responsible department for any kind to lodge a complaint or seek help. 
Secondly, the people lodge the complaints but cannot get the details of the problem. 
This system does not have much popularity and is not user friendly.
DRAW BACKS 
Inconsistency in customer interaction 
Corruption
Lack of prompt updating as to when the complaint or issues will be resolved.
All records are maintained on paper documents which makes it less secure as it can be misplaced or misused. 
Complaints are not being registered to the proper departments.
Members do not have channel for tracking their complaints.

SCOPE
The system is computerised to do all the work.
Complaints are easier to monitor, track and  co-ordinate with respective people and resolve.
Controlling redundancy in storing the data. 
User friendly and data remains confidential.
This system saves time and eradicates corruption.

DATA REQUIREMENTS 
Login
Uploading details
If any issue post complaint in the online portal 
Will be checked by the admin of society
Meeting will be held and reminder will be given
Contacting maintenance person